[feature] move task relations into own concern trait
