+++ app/Models/Task.php
@@ -2,12 +2,14 @@
 
 namespace App\Models;
 
+use App\Models\Task\Concerns;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Task extends Model
 {
     use HasFactory;
+    use Concerns\HasRelations;
 
     /**
      * The attributes that are mass assignable.
@@ -22,34 +24,4 @@ class Task extends Model
         'task',
         'type',
     ];
-
-    public function creator() {
-        return $this->belongsTo(
-            User::class,
-            'creator_id'
-        );
-    }
-    public function assigned()
-	{
-        return  $this->belongsTo(User::class, 'assignedto', 'id');
-        
-    }
-
-    public function items() {
-        return $this->hasMany(
-            TaskItem::class
-        );
-    }
-
-    public function updates() {
-        return $this->hasMany(
-            ProjectUpdate::class
-        );
-    }
-
-    public function project() {
-        return $this->belongsTo(
-            Project::class
-        );
-    }
 }
